+++ client/src/test/java/replicant/TransportContextImplTest.java
@@ -52,7 +52,7 @@ public class TransportContextImplTest
 
     final TestSpyEventHandler handler = registerTestSpyEventHandler();
     handler.assertEventCount( 0 );
-    context.onError( new Throwable() );
+    context.onError();
     handler.assertEventCount( 2 );
 
     handler.assertEvent( MessageReadFailureEvent.class );
@@ -67,7 +67,7 @@ public class TransportContextImplTest
     handler.assertEventCount( 2 );
 
     // This should be ignored
-    context.onError( new Throwable() );
+    context.onError();
 
     handler.assertEventCount( 2 );
   }