+++ client/src/main/java/replicant/Scheduler.java
@@ -1,6 +1,8 @@
 package replicant;
 
 import elemental2.dom.DomGlobal;
+import java.util.Timer;
+import java.util.TimerTask;
 import javax.annotation.Nonnull;
 
 /**
@@ -13,6 +15,11 @@ final class Scheduler
   static void schedule( @Nonnull final SafeFunction<Boolean> command )
   {
     c_support.schedule( command );
+  }
+
+  static void scheduleOnceOff( @Nonnull final SafeProcedure command, final int delay )
+  {
+    c_support.scheduleOnceOff( command, delay );
   }
 
   /**
@@ -30,10 +37,31 @@ final class Scheduler
       {
       }
     }
+
+    @GwtIncompatible
+    @Override
+    void scheduleOnceOff( @Nonnull final SafeProcedure command, final int delay )
+    {
+      final TimerTask timerTask = new TimerTask()
+      {
+        @Override
+        public void run()
+        {
+          command.call();
+        }
+      };
+      final Timer timer = new Timer();
+      timer.schedule( timerTask, delay );
+    }
   }
 
   private static abstract class AbstractSchedulerSupport
   {
+    void scheduleOnceOff( @Nonnull final SafeProcedure command, final int delay )
+    {
+      DomGlobal.setTimeout( v -> command.call(), delay );
+    }
+
     void schedule( @Nonnull final SafeFunction<Boolean> command )
     {
       final long end = System.currentTimeMillis() + 14;
@@ -46,25 +74,6 @@ final class Scheduler
       }
       DomGlobal.setTimeout( v -> schedule( command ) );
     }
-/*
-    When we no longer need to support IE11 we should move to this code.
-
-    void schedule( @Nonnull final SafeFunction<Boolean> command )
-    {
-      DomGlobal.requestIdleCallback( deadline -> idleSchedule( deadline, command ) );
-    }
-
-    private void idleSchedule( @Nonnull final IdleDeadline deadline, @Nonnull final SafeFunction<Boolean> command )
-    {
-      while ( !deadline.isDidTimeout() )
-      {
-        if ( !command.call() )
-        {
-          return;
-        }
-      }
-      DomGlobal.requestIdleCallback( v -> idleSchedule( v, command ) );
-*/
   }
 
   private Scheduler()