+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v6.40](https://github.com/replicant4j/replicant/tree/v6.40) (2019-04-17)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.39...v6.40)